OpenWebUI/mcp/實作範例/mcpo-sql/mcpo/requirements.txt
mcpo
mcp
psycopg2-binary